chore: added CI/CD , cleaned up README.md, added badges  (#1)
* Added Jenkinsfile

* Added Jenkinsfile

* Added Jenkinsfile

* Added Jenkinsfile

* Added Jenkinsfile

* cicd pipeline

* Test and package step

* correct configs

* setup tests

* use proper https for unit tests in CI CD pipeline

* chore: added Jenkins Build

* use longer file for upload CI test

* bigger file

* dont use chunked upload for unit test

* jenkins formated tests

* fancy reports

* fancy reports

* Added Jenkinsfile

* build badge

* build badge

* small fixes in the readme.md

* chore:  added badges ci/cd npm, downloads

+++ test/mindconnect-setup.spec.ts
@@ -74,7 +74,7 @@ describe("MindConnect Setup", () => {
 
     it("should register maximum 5 agents for diagnostics", mochaAsync(async () => {
 
-        if (!process.env.TENANT) {
+        if (!process.env.CI) {
             // don't delete all diagnostic registrations all the time on the CI/CD this can disturb the normal workings on the tenants.
             const mcsetup = new MindConnectSetup(gateway, basicAuth, tenant);
             mcsetup.should.not.be.null;
@@ -82,12 +82,8 @@ describe("MindConnect Setup", () => {
             await mcsetup.DeleteAllDiagnosticActivations();
 
             try {
-                await mcsetup.RegisterForDiagnostic("00000000000000000000000000000000");
-                await mcsetup.RegisterForDiagnostic("00000000000000000000000000000001");
-                await mcsetup.RegisterForDiagnostic("00000000000000000000000000000002");
-                await mcsetup.RegisterForDiagnostic("00000000000000000000000000000003");
-                await mcsetup.RegisterForDiagnostic("00000000000000000000000000000004");
-                await mcsetup.RegisterForDiagnostic("00000000000000000000000000000005");
+                await mcsetup.RegisterForDiagnostic("2903bf15381646d3a8f4aeeff8d9bd29");
+                await mcsetup.RegisterForDiagnostic("68766a93af834984a8f8decfbeec961e");
             } catch (err) {
                 if ((("" + err).indexOf("agent limitation")) < 0) {
                     throw err;
@@ -102,28 +98,32 @@ describe("MindConnect Setup", () => {
     }));
 
     it("should get logs on error", mochaAsync(async () => {
-        const mcsetup = new MindConnectSetup(gateway, basicAuth, tenant);
-        mcsetup.should.not.be.null;
-        const agent = new MindConnectAgent(sharedSecretConfig);
-        agent.should.exist;
+        if (!process.env.CI) {
+            const mcsetup = new MindConnectSetup(gateway, basicAuth, tenant);
+            mcsetup.should.not.be.null;
+            const agent = new MindConnectAgent(sharedSecretConfig);
+            agent.should.exist;
 
-        try {
-            const response = await mcsetup.RegisterForDiagnostic(agent.ClientId());
-        } catch (err) {
-            if ((("" + err).indexOf("Conflict")) < 0) {
-                throw err;
+            await mcsetup.DeleteAllDiagnosticActivations();
+
+            try {
+                const response = await mcsetup.RegisterForDiagnostic(agent.ClientId());
+            } catch (err) {
+                if ((("" + err).indexOf("Conflict")) < 0) {
+                    throw err;
+                }
             }
+            if (!agent.IsOnBoarded()) {
+                agent.OnBoard();
+            }
+            await agent.PostData([{ dataPointId: "Unexistent", qualityCode: "123123135", value: "12312346.42.23" }], undefined, false);
+            const diag = await mcsetup.GetDiagnosticInformation(agent.ClientId());
+            diag.content.should.exist;
+            let activations = await mcsetup.GetDiagnosticActivations();
+            activations.content.length.should.be.greaterThan(0);
+            await mcsetup.DeleteDiagnostic(agent.ClientId());
+            activations = await mcsetup.GetDiagnosticActivations();
+            activations.content.length.should.be.equal(0);
         }
-        if (!agent.IsOnBoarded()) {
-            agent.OnBoard();
-        }
-        await agent.PostData([{ dataPointId: "Unexistent", qualityCode: "123123135", value: "12312346.42.23" }], undefined, false);
-        const diag = await mcsetup.GetDiagnosticInformation(agent.ClientId());
-        diag.content.should.exist;
-        let activations = await mcsetup.GetDiagnosticActivations();
-        activations.content.length.should.be.greaterThan(0);
-        await mcsetup.DeleteDiagnostic(agent.ClientId());
-        activations = await mcsetup.GetDiagnosticActivations();
-        activations.content.length.should.be.equal(0);
     }));
 });
